requirements.txt
google.generativeai
Pillow
fastapi
json
os
ultralytics